+++ CrossCutting/Authorization/IJwtTokenGenerator.cs
@@ -1,0 +1,7 @@
+﻿namespace CrossCutting.Authorization
+{
+    public interface IJwtTokenGenerator
+    {
+        string CreateJwtToken();
+    }
+}
