+++ DopplerJobsServer/appsettings.QA.json
@@ -6,7 +6,8 @@
     "HolidayRetryCountLimit": 5
   },
   "DopplerSapConfiguration": {
-    "BusinessPartnerUrl": "https://192.168.6.35:65100/BusinessPartner/CreateOrUpdateBusinessPartner"
+    "CurrencyEndpoint": "https://apisqa.fromdoppler.net/sap/Billing/SetCurrencyRate",
+    "BillingEndpoint": "https://apisqa.fromdoppler.net/sap/Billing/CreateBillingRequest"
   },
   "Jobs": {
     "DopplerBillingJobSettings": {
