+++ DopplerJobsServer/appsettings.json
@@ -50,14 +50,14 @@
     }
   },
   "DopplerCurrencyServiceSettings": {
-    "Url": "https://localhost:44353/",
+    "Url": "http://localhost:5001/conversion/",
     "CurrencyCodeList": [ "ARS" ],
     "InsertCurrencyQuery": "[dbo].[InsertNewCurrencyRate]",
     "HolidayRetryCountLimit": 5
   },
   "DopplerSapConfiguration": {
-    "CurrencyEndpoint": "http://172.24.16.82:10550/Billing/SetCurrencyRate",
-    "BillingEndpoint": "http://172.24.16.82:10550/Billing/CreateBillingRequest"
+    "CurrencyEndpoint": "http://localhost:5002/Billing/SetCurrencyRate",
+    "BillingEndpoint": "http://localhost:5002/Billing/CreateBillingRequest"
   },
   "ConnectionStrings": {
     "DopplerDatabase": "[SECRET_KEY]"
