+++ DopplerJobsServer/appsettings.prod.json
@@ -6,8 +6,8 @@
     "HolidayRetryCountLimit": 5
   },
   "DopplerSapConfiguration": {
-    "CurrencyEndpoint": "http://172.24.16.82:65100/Billing/SetCurrencyRate",
-    "BillingEndpoint": "http://172.24.16.82:65100/Billing/CreateBillingRequest"
+    "CurrencyEndpoint": "http://sap/Billing/SetCurrencyRate",
+    "BillingEndpoint": "http://sap/Billing/CreateBillingRequest"
   },
   "Jobs": {
     "DopplerBillingJobSettings": {
